+++ src/MemoryDump.cpp
@@ -1,5 +1,7 @@
 #include "MemoryDump.h"
 
+#include <RED4ext/Scripting/Natives/Generated/ent/IComponent.hpp>
+
 #include "MemoryTargetAddress.h"
 #include "MemoryTargetWeakHandle.h"
 
@@ -20,10 +22,10 @@ Red::Handle<MemoryTarget> MemoryDump::track_address(
   const Red::CString& p_name, const Red::CName& p_type, uint64_t p_address,
   const Red::Optional<uint32_t, 0x38>& p_size) {
   auto target = Red::MakeHandle<MemoryTargetAddress>(
-    p_name, p_type, reinterpret_cast<uint8_t*>(p_address), p_size.value);
+      p_name, p_type, reinterpret_cast<uint8_t*>(p_address), p_size.value);
 
   target->capture();
   return target;
 }
 
-}  // namespace RedMemoryDump
+} // namespace RedMemoryDump